update
update

+++ workspace/springbootdemo/src/main/java/com/loveGod/demo/model/RegisterDao.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 //import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface RegisterDao extends JpaRepository<RegisterModel, Long> {
+public interface RegisterDao extends JpaRepository<RegisterModel, Integer> {
 
 	// ================================== 使用者登入判斷帳密是否正確 ========================
 	@Query("select r from RegisterModel r where r.memberId = ?1 and r.password = ?2 ")
@@ -14,10 +14,4 @@ public interface RegisterDao extends JpaRepository<RegisterModel, Long> {
 
 	@Query("select r from RegisterModel r where r.memberId = ?1 ")
 	List<RegisterModel> findMemberId(String memberId);
-
-//	// 業務邏輯
-//		public MemberAccountVO login(MemberAccount memberAccount);
-//		public Optional<String> register(MemberAccountVO memberAccountVO);
-//	RegisterModel findRegisterModelByMemberId(String memberId);
-
 }
